+++ CHANGELOG.md
@@ -4,13 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Pre-Alpha 4 [8/13/18]
+### Additions
+- Added Missing Polygon Functions.
+
+### Fixes
+- Fixed Circle circumferenceByRadius()
+
 ## Pre-Alpha 3 [7/31/18]
+### Additions
 - ADDED: Basic functionality for cubes.
 - ADDED: Empty classes for soon to be working shapes.
+
+### Fixes
 - FIXED: perimeterByArea() in Parallelogram instead of Square.
 
 ## Pre-Alpha 2 [7/26/18]
+### Additions
 - ADDED: Essential functionality for three sided shapes and four sided shapes.
 
 ## Pre-Alpha 1 [7/12/18]
+### Additions
 - ADDED: Basic functionality for rectangles, circles, triangles, and lines.